refactor: move the lake coupling under advanced_packages
An advanced package carries its own dependent variable and needs its own
equation in the bordered adjoint system, so give them a home rather than adding
a module at the top level for each. sfr6 follows here.
